removed argon2's dependency by separating codes using argon2 into different file
