+++ .travis.yml
@@ -6,7 +6,6 @@ python:
 sudo: false
 install:
   - pip install sqlalchemy
-  - pip install argon2-cffi
   - pip install pytest
   - python setup.py install
 script:
