+++ quesadiya/django_tool/database/__init__.py
@@ -1,1 +1,2 @@
-from .django_database import *
+from .django_database import create_django_tables
+from .insert_utils import insert_admin, insert_collaborator
